Add new packages to organise classes
