+++ src/main/java/seedu/duke/financial/Expense.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.financial;
 
 /**
  * Represents an expense transaction.
