+++ src/main/java/seedu/duke/financial/FinancialList.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.financial;
 
 import java.util.ArrayList;
