+++ src/main/java/seedu/duke/financial/Income.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.financial;
 
 /**
  * Represents an income transaction.
